mfd: zl80x32: Add supported firmware images info
Add info about supported firmware hex-images.The hardware currently
supports 11 types:

* Utility
* Firmware 1..3
* Config 0..6

The Utility is special, it is loaded by the driver into HW internal
memory and then it is executed. The utility acts as helper for
the driver and provides flash capability on the firmware side.

Other images is loaded by the driver into specific HW internal memory
location and then the driver send a flash request to flash utility.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>