Sample module to load zl3073x dpll driver without DT
This is very simple sample module to show how to create zl3073x
device instance on non-DT platform.

The sample uses software_node to specify device properties and
optionally PHC devices. It also shows how to pass platform data
to zl3073x driver with PIN properties instead of hardcoding them
into driver.

The sample is created for Microchip EDS2 board so to use it in
different environment one has to provide a method to access
to the I2C adapter where the DPLL chip is connected to.

This information could be provided maybe by ACPI or DMI. In any
case the following has to be satisfied:
1. Target platform has to be able to identify if the DPLL chip
   is present or not.
2. If it is present the target has to know what I2C bus is used
   for the DPLL chip
3. If the chip is directly connected to physical bus then
   the i2c_adapter associated with this bus should be used.
4. If the chip is connected to I2C mux on some physical bus then
   platform has to know what channel of the mux is used for the
   DPLL. If so the I2C adapter associated with this I2C mux channel
   should be used.

TODO:
Driver will be changed so the PIN properties will be passed
using software_node properties instead of platform_data. This
will allow to pass the board specific properties through
device tree on DT platforms or through software_node on non-DT
platforms.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>