dpll: zl3073x: Fix phase offsets
The phase_offset returned to DPLL core uses ps units but it has
3-digit decimal fractional part, this is in practice fs units.
The device reports phase offsets in units 0.01 ps so we can
multiply the value by 10 to convert it to fs.

Multiply the reported phase offset by 10 instead of division by 100
and adjust the rest of computation to use fs instead of ps units.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>